+++ UserInputOutput.h
@@ -3,10 +3,10 @@
 
 struct UserInput
 {
-	unsigned iniInvestAmount = 0;
-	unsigned monthlyDeposit = 0;
-	unsigned annualIntrestRate = 0;
-	unsigned short numOfYears = 0;
+	double iniInvestAmount = 0.0;
+	double monthlyDeposit = 0.0;
+	double annualIntrestRate = 0.0;
+	unsigned short numOfYears = 0.0;
 };
 
 class UserInputOutput
